refactor: Re-organize source files
